requirements.txt
# MetaFusion Requirements
psutil
plexapi
ruamel.yaml
orjson
pycountry
Pillow
aiohttp
aiolimiter
tqdm
